requirements.txt
aiogram==3.15.0
apscheduler==3.10.4
python-dotenv==1.0.1
